[MIG] announcement: Migration to 15.0
TT40347
