+++ announcement/i18n/announcement.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 13.0\n"
+"Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -26,7 +26,7 @@ msgstr ""
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #, python-format
 msgid "Activities"
 msgstr ""
@@ -48,7 +48,7 @@ msgstr ""
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #: model:ir.model.fields,field_description:announcement.field_announcement_log__announcement_id
 #: model:ir.model.fields,field_description:announcement.field_read_announcement_wizard__announcement_id
 #: model:ir.module.category,name:announcement.category_announcement
@@ -58,14 +58,12 @@ msgid "Announcement"
 msgstr ""
 
 #. module: announcement
-#: model:ir.actions.act_window,name:announcement.action_announcement_log
 #: model:ir.model.fields,field_description:announcement.field_announcement__announcement_log_ids
 msgid "Announcement Log"
 msgstr ""
 
 #. module: announcement
-#: model_terms:ir.ui.view,arch_db:announcement.announcement_log_view_tree
-#: model_terms:ir.ui.view,arch_db:announcement.read_announcement_wizard_view_tree
+#: model:ir.actions.act_window,name:announcement.action_announcement_log
 msgid "Announcement Logs"
 msgstr ""
 
@@ -81,10 +79,9 @@ msgstr ""
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #: model:ir.actions.act_window,name:announcement.announcement_action
 #: model:ir.ui.menu,name:announcement.menu_announcement_management
-#: model_terms:ir.ui.view,arch_db:announcement.announcement_view_tree
 #, python-format
 msgid "Announcements"
 msgstr ""
@@ -166,14 +163,14 @@ msgstr ""
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/js/announcement.js:0
+#: code:addons/announcement/static/src/js/announcement_systray.esm.js:0
 #, python-format
 msgid "Mark as read"
 msgstr ""
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #, python-format
 msgid "No announcements."
 msgstr ""
@@ -206,6 +203,12 @@ msgstr ""
 #. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_read_announcement_wizard__date
 msgid "Read Date"
+msgstr ""
+
+#. module: announcement
+#: code:addons/announcement/models/announcement.py:0
+#, python-format
+msgid "Read Logs"
 msgstr ""
 
 #. module: announcement
