+++ announcement/i18n/es.po
@@ -4,10 +4,10 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 13.0\n"
+"Project-Id-Version: Odoo Server 15.0\n"
 "Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2022-08-22 11:10+0000\n"
-"PO-Revision-Date: 2022-08-22 13:13+0200\n"
+"POT-Creation-Date: 2022-11-18 09:12+0000\n"
+"PO-Revision-Date: 2022-11-18 10:13+0100\n"
 "Last-Translator: \n"
 "Language-Team: \n"
 "Language: es\n"
@@ -15,7 +15,7 @@ msgstr ""
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: \n"
-"X-Generator: Poedit 2.3\n"
+"X-Generator: Poedit 3.1.1\n"
 
 #. module: announcement
 #: model_terms:ir.ui.view,arch_db:announcement.announcement_management_view_form
@@ -30,7 +30,7 @@ msgstr "Activo"
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #, python-format
 msgid "Activities"
 msgstr "Actividades"
@@ -52,7 +52,7 @@ msgstr "Anunciar a"
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #: model:ir.model.fields,field_description:announcement.field_announcement_log__announcement_id
 #: model:ir.model.fields,field_description:announcement.field_read_announcement_wizard__announcement_id
 #: model:ir.module.category,name:announcement.category_announcement
@@ -62,16 +62,14 @@ msgid "Announcement"
 msgstr "Anuncio"
 
 #. module: announcement
-#: model:ir.actions.act_window,name:announcement.action_announcement_log
 #: model:ir.model.fields,field_description:announcement.field_announcement__announcement_log_ids
 msgid "Announcement Log"
 msgstr "Registro del anuncio"
 
 #. module: announcement
-#: model_terms:ir.ui.view,arch_db:announcement.announcement_log_view_tree
-#: model_terms:ir.ui.view,arch_db:announcement.read_announcement_wizard_view_tree
+#: model:ir.actions.act_window,name:announcement.action_announcement_log
 msgid "Announcement Logs"
-msgstr "Registros de anuncio"
+msgstr "Registro del anuncio"
 
 #. module: announcement
 #: model:res.groups,name:announcement.announcemenent_manager
@@ -85,10 +83,9 @@ msgstr "Tipo de anuncio"
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #: model:ir.actions.act_window,name:announcement.announcement_action
 #: model:ir.ui.menu,name:announcement.menu_announcement_management
-#: model_terms:ir.ui.view,arch_db:announcement.announcement_view_tree
 #, python-format
 msgid "Announcements"
 msgstr "Anuncios"
@@ -140,7 +137,7 @@ msgstr "ID (identificación)"
 #. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_announcement__in_date
 msgid "In Date"
-msgstr ""
+msgstr "En fecha"
 
 #. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_announcement____last_update
@@ -170,14 +167,14 @@ msgstr "Registrar lecturas de usuario"
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/js/announcement.js:0
+#: code:addons/announcement/static/src/js/announcement_systray.esm.js:0
 #, python-format
 msgid "Mark as read"
 msgstr "Marcar como léido"
 
 #. module: announcement
 #. openerp-web
-#: code:addons/announcement/static/src/xml/announcement.xml:0
+#: code:addons/announcement/static/src/xml/announcement_systray.xml:0
 #, python-format
 msgid "No announcements."
 msgstr "No hay anuncios."
@@ -213,6 +210,12 @@ msgid "Read Date"
 msgstr "Fecha de lectura"
 
 #. module: announcement
+#: code:addons/announcement/models/announcement.py:0
+#, python-format
+msgid "Read Logs"
+msgstr "Registros de lectura"
+
+#. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_read_announcement_wizard__read_state
 msgid "Read State"
 msgstr "Estado de lectura"
@@ -245,7 +248,7 @@ msgstr "Usuarios específicos"
 #. module: announcement
 #: model:ir.model,name:announcement.model_ir_config_parameter
 msgid "System Parameter"
-msgstr ""
+msgstr "Parámetro del sistema"
 
 #. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_announcement__name
@@ -270,7 +273,7 @@ msgstr "Anuncio pendiente"
 #. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_read_announcement_wizard__user_id
 msgid "User"
-msgstr "Usuario\t"
+msgstr "Usuario"
 
 #. module: announcement
 #: model:ir.model.fields,field_description:announcement.field_announcement__user_group_ids
@@ -307,6 +310,3 @@ msgstr "Válido hasta"
 #: model_terms:ir.ui.view,arch_db:announcement.announcement_view_form
 msgid "e.g. Announcement description..."
 msgstr "p.e. Descripción del anuncio..."
-
-#~ msgid "HTTP Routing"
-#~ msgstr "Ruta HTTP"
